requirements.txt
joblib
requests
beautifulsoup4
pytesseract
Pillow
streamlit
scikit-learn
langdetect
deep-translator
pandas
numpy
scipy

